pyproject.toml
[project]
name = "nb-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arclet-cithun>=0.1.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.54.2",
    "nonebot-plugin-htmlrender>=0.5.1",
    "nonebot-plugin-manager",
    "nonebot-plugin-mmm>=0.2.3",
    "nonebot-plugin-pong>=0.1.2",
    "nonebot2[fastapi,httpx]>=2.4.1",
]

[tool.uv.sources]
nonebot-plugin-manager = { git = "https://github.com/nonepkg/plugin-manager" }
